+++ playground/angular/src/app/app.module.ts
@@ -2,15 +2,29 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { SwiperModule } from 'src/angular/src/public-api';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule.withServerTransition({ appId: 'serverApp' }),
-    SwiperModule,
     BrowserAnimationsModule,
+    RouterModule.forRoot([
+      {
+        path: '',
+        redirectTo: 'home',
+        pathMatch: 'full',
+      },
+      {
+        path: 'home',
+        loadChildren: () => import('./home/home.module').then((m) => m.HomeModule),
+      },
+      {
+        path: 'details',
+        loadChildren: () => import('./detail/detail.module').then((m) => m.DetailModule),
+      },
+    ]),
   ],
   providers: [],
   bootstrap: [AppComponent],