+++ src/components/core/slide/slideNext.js
@@ -1,7 +1,8 @@
 /* eslint no-unused-vars: "off" */
 export default function slideNext(speed = this.params.speed, runCallbacks = true, internal) {
   const swiper = this;
-  const { params, animating } = swiper;
+  const { params, animating, enabled } = swiper;
+  if (!enabled) return swiper;
   const increment = swiper.activeIndex < params.slidesPerGroupSkip ? 1 : params.slidesPerGroup;
   if (params.loop) {
     if (animating && params.loopPreventsSlide) return false;